Simplify circle (#8)
I pulled out all of the docker deps in the circleci builds, because to
run them requires access to our private docker registry. This is also a
better direction as it makes what we are doing in circle clearer,
pulling commands which are invoked through the base image out into
explicit commands in `circle.yml`.